requirements.txt
aiohttp==3.8.1
asgiref==3.5.2
astroid==2.12.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.0
dicttoxml==1.7.4
Django==3.2.4
django-environ==0.7.0
django-ipware==4.0.0
django-livereload-server==0.4
django-maintenance-mode==0.16.1
docutils==0.19
future==0.18.2
geocoder==1.38.1
geoip2==4.4.0
idna==3.3
importlib-metadata==4.8.1
ip2geotools==0.1.6
IP2Location==8.6.2
isort==5.9.3
jeepney==0.7.1
keyring==23.2.1
lazy-object-proxy==1.7.1
lxml==4.9.1
maxminddb==2.2.0
mccabe==0.7.0
multidict==5.1.0
mysqlclient==2.1.1
packaging==21.0
pip-review==1.1.0
pkginfo==1.7.1
platformdirs==2.5.2
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
pylint==2.11.1
pyparsing==2.4.7
pyquery==1.4.3
pytz==2021.1
ratelim==0.1.6
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
selenium==4.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
tqdm==4.62.3
twine==3.4.2
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==4.3.0
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.12.1
yarl==1.7.2
zipp==3.8.1
django-taggit==3.0.0